+++ examples/blaze/src/main/scala/com/example/http4s/blaze/BlazeMetricsExample.scala
@@ -15,7 +15,7 @@ object BlazeMetricsExample extends StreamApp {
     "/metrics" -> metricsService(metricRegistry)
   )
 
-  def main(args: List[String]) = BlazeBuilder.bindHttp(8080)
+  def stream(args: List[String]) = BlazeBuilder.bindHttp(8080)
     .mountService(srvc, "/http4s")
     .serve
 }